refactor onboarding to extract country/city to component
